# Increase `open:` query performance on large repositories