--- a/server/bleep/benches/indexes.rs
+++ b/server/bleep/benches/indexes.rs
@@ -71,7 +71,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         b.iter_batched(
             || {
                 doc! {
-                    file.file_disk_path => "js-sample-big-symbols.js",
+                    file.entry_disk_path => "js-sample-big-symbols.js",
                     file.repo_ref => "local//bloop",
                     file.repo_name => "bloop",
                     file.relative_path => "js-sample-big-symbols.js",