--- a/server/bleep/src/indexes/file.rs
+++ b/server/bleep/src/indexes/file.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::{HashMap, HashSet},
     ops::Not,
-    path::{Path, PathBuf},
+    path::{Path, PathBuf, MAIN_SEPARATOR},
     sync::Arc,
 };
 
@@ -44,7 +44,7 @@ use crate::{
 };
 
 struct Workload<'a> {
-    file_disk_path: PathBuf,
+    entry_disk_path: PathBuf,
     repo_disk_path: &'a Path,
     repo_ref: String,
     repo_name: &'a str,
@@ -61,8 +61,8 @@ pub struct File {
     #[cfg(feature = "debug")]
     histogram: Arc<RwLock<Histogram>>,
 
-    // Path to the indexed file on disk
-    pub file_disk_path: Field,
+    // Path to the indexed file or directory on disk
+    pub entry_disk_path: Field,
     // Path to the root of the repo on disk
     pub repo_disk_path: Field,
     // Path to the file, relative to the repo root
@@ -105,7 +105,7 @@ impl File {
                 .set_index_option(IndexRecordOption::WithFreqsAndPositions),
         );
 
-        let file_disk_path = builder.add_text_field("file_disk_path", STRING);
+        let entry_disk_path = builder.add_text_field("entry_disk_path", STRING);
         let repo_disk_path = builder.add_text_field("repo_disk_path", STRING);
         let repo_ref = builder.add_text_field("repo_ref", STRING | STORED);
         let repo_name = builder.add_text_field("repo_name", trigram.clone());
@@ -131,7 +131,7 @@ impl File {
         let raw_relative_path = builder.add_bytes_field("raw_relative_path", FAST);
 
         Self {
-            file_disk_path,
+            entry_disk_path,
             repo_disk_path,
             relative_path,
             repo_ref,
@@ -241,8 +241,6 @@ impl Indexable for File {
                     None
                 }
             })
-            // Ignore directories from walk result.
-            .filter(|de| matches!(de.file_type(), Some(ft) if ft.is_file()))
             // Preliminarily ignore files that are very large, without reading the contents.
             .filter(|de| matches!(de.metadata(), Ok(meta) if meta.len() < MAX_FILE_LEN))
             .map(|de| crate::canonicalize(de.into_path()).unwrap())
@@ -252,28 +250,31 @@ impl Indexable for File {
         let start = std::time::Instant::now();
 
         use rayon::prelude::*;
-        walker.par_iter().for_each(|file_disk_path| {
-            let workload = Workload {
-                file_disk_path: file_disk_path.clone(),
-                repo_disk_path: &repo.disk_path,
-                repo_ref: reporef.to_string(),
-                repo_name: &repo_name,
-                cache: &file_cache,
-                repo_info,
-            };
+        walker
+            .into_par_iter()
+            .chain(Some(repo.disk_path.clone()))
+            .for_each(|entry_disk_path| {
+                let workload = Workload {
+                    entry_disk_path: entry_disk_path.clone(),
+                    repo_disk_path: &repo.disk_path,
+                    repo_ref: reporef.to_string(),
+                    repo_name: &repo_name,
+                    cache: &file_cache,
+                    repo_info,
+                };
 
-            debug!(?file_disk_path, "queueing file");
-            if let Err(err) = self.worker(workload, writer) {
-                warn!(%err, ?file_disk_path, "indexing failed; skipping");
-            }
-        });
+                debug!(?entry_disk_path, "queueing entry");
+                if let Err(err) = self.worker(workload, writer) {
+                    warn!(%err, ?entry_disk_path, "indexing failed; skipping");
+                }
+            });
 
         info!(?repo.disk_path, "file indexing finished, took {:?}", start.elapsed());
 
         file_cache.retain(|k, v| {
             if v.fresh.not() {
                 writer.delete_term(Term::from_field_text(
-                    self.file_disk_path,
+                    self.entry_disk_path,
                     &k.to_string_lossy(),
                 ));
             }
@@ -308,7 +309,7 @@ impl Indexer<File> {
         let searcher = reader.searcher();
 
         let query = TermQuery::new(
-            Term::from_field_text(self.source.file_disk_path, file_disk_path),
+            Term::from_field_text(self.source.entry_disk_path, file_disk_path),
             IndexRecordOption::Basic,
         );
 
@@ -429,10 +430,10 @@ impl Indexer<File> {
 }
 
 impl File {
-    #[tracing::instrument(fields(repo=%workload.repo_ref, file_path=?workload.file_disk_path), skip_all)]
+    #[tracing::instrument(fields(repo=%workload.repo_ref, entry_disk_path=?workload.entry_disk_path), skip_all)]
     fn worker(&self, workload: Workload<'_>, writer: &IndexWriter) -> Result<()> {
         let Workload {
-            file_disk_path,
+            entry_disk_path,
             repo_ref,
             repo_disk_path,
             repo_name,
@@ -443,15 +444,25 @@ impl File {
         #[cfg(feature = "debug")]
         let start = Instant::now();
 
-        let mut buffer = match std::fs::read_to_string(&file_disk_path) {
-            Err(err) => {
-                debug!(%err, ?file_disk_path, "read failed; skipping");
-                return Ok(());
+        let mut buffer = if entry_disk_path.is_file() {
+            match std::fs::read_to_string(&entry_disk_path) {
+                Err(err) => {
+                    debug!(%err, ?entry_disk_path, "read failed; skipping");
+                    return Ok(());
+                }
+                Ok(buffer) => buffer,
             }
-            Ok(buffer) => buffer,
+        } else {
+            String::new()
         };
 
-        let relative_path = file_disk_path.strip_prefix(repo_disk_path)?;
+        let relative_path = entry_disk_path.strip_prefix(repo_disk_path)?;
+        let relative_path_str = if entry_disk_path.is_dir() {
+            format!("{}{MAIN_SEPARATOR}", relative_path.to_string_lossy()).into()
+        } else {
+            relative_path.to_string_lossy()
+        };
+
         trace!("processing file");
 
         let content_hash = {
@@ -463,7 +474,7 @@ impl File {
 
         trace!("adding cache entry");
 
-        match cache.entry(file_disk_path.clone()) {
+        match cache.entry(entry_disk_path.clone()) {
             Entry::Occupied(mut val) if val.get().value == content_hash => {
                 // skip processing if contents are up-to-date in the cache
                 val.get_mut().fresh = true;
@@ -481,7 +492,7 @@ impl File {
         let lang_str = repo_info
             .langs
             .path_map
-            .get(&file_disk_path)
+            .get(&entry_disk_path)
             .unwrap_or_else(|| {
                 warn!("Path not found in language map");
                 &Some("")
@@ -504,7 +515,7 @@ impl File {
                         Some(syms) => SymbolLocations::Ctags(syms.clone()),
                         // no ctags either
                         _ => {
-                            debug!(%lang_str, ?file_disk_path, "failed to build tags");
+                            debug!(%lang_str, ?entry_disk_path, "failed to build tags");
                             SymbolLocations::Empty
                         }
                     }
@@ -558,8 +569,8 @@ impl File {
         let buf_size = buffer.len();
         writer.add_document(doc!(
             self.repo_disk_path => repo_disk_path.to_string_lossy().as_ref(),
-            self.file_disk_path => file_disk_path.to_string_lossy().as_ref(),
-            self.relative_path => relative_path.to_string_lossy().as_ref(),
+            self.entry_disk_path => entry_disk_path.to_string_lossy().as_ref(),
+            self.relative_path => relative_path_str.as_ref(),
             self.repo_ref => repo_ref,
             self.repo_name => repo_name,
             self.content => buffer.as_str(),
@@ -571,7 +582,7 @@ impl File {
             self.symbols => symbols,
             self.raw_content => buffer.as_bytes(),
             self.raw_repo_name => repo_name.as_bytes(),
-            self.raw_relative_path => relative_path.to_string_lossy().as_ref().as_bytes(),
+            self.raw_relative_path => relative_path_str.as_ref().as_bytes(),
         ))?;
 
         trace!("document written");