With the `tensorflow` repository, this takes `open:` query time down from ~1 second to ~20 milliseconds, a 50x speedup.

The central idea behind this change involves introducing dummy documents into the file index, representing directories. This allows us to search for folders, without having to infer their existence due to other documents.

Before, 1.028 seconds:
```
curl -v 'localhost:7878/q?q=open:true%20repo:tensorflow'  0.00s user 0.00s system 10% cpu 1.028 total
```

After, 19 milliseconds:
```
curl -v 'localhost:7878/q?q=open:true%20repo:tensorflow'  0.00s user 0.00s system 47% cpu 0.019 total
```

Additionally, a bug with open queries of the form `some/valid/file/with/trailing/slash.txt/` resulting in panics was fixed.